+++ packages/gestalt-codemods/0.108.0-0.109.0/__testfixtures__/convert-font-weight-dynamic.output.js
@@ -1,0 +1,15 @@
+// @flow
+import React from 'react';
+import { Text } from 'gestalt';
+
+export default function BoldText({
+  isCurrentPage,
+}: {
+  isCurrentPage?: boolean,
+}) {
+  return (
+    <Text size="xl" weight={isCurrentPage ? "bold" : "normal"}>
+      Bold
+    </Text>
+  );
+}